@@ -14,8 +14,8 @@
     "framer-motion": "^6.2.8",
     "jwt-decode": "^3.1.2",
     "query-string": "^7.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-facebook-login": "^4.1.1",
     "react-fluent-form": "^1.6.0",
     "react-google-login": "^5.2.2",
